refactor: mapを直接記述
